+++ latch_cli/snakemake/serialize.py
@@ -323,6 +323,8 @@ def generate_snakemake_entrypoint(
         from typing import NamedTuple
         import stat
         import sys
+        from dataclasses import is_dataclass, asdict
+        from enum import Enum
 
         from flytekit.extras.persistence import LatchPersistence
         import traceback
@@ -333,6 +335,18 @@ def generate_snakemake_entrypoint(
 
         sys.stdout.reconfigure(line_buffering=True)
         sys.stderr.reconfigure(line_buffering=True)
+
+        def get_parameter_json_value(v):
+            if is_dataclass(v):
+                return asdict(v)
+            elif isinstance(v, Enum):
+                return v.value
+            elif isinstance(v, list):
+                return [get_parameter_json_value(x) for x in v]
+            elif isinstance(v, dict):
+                return {k: get_parameter_json_value(x) for k, x in v.items()}
+            else:
+                return v
 
         def check_exists_and_rename(old: Path, new: Path):
             if new.exists():
@@ -395,6 +409,8 @@ def generate_jit_register_code(
         from typing import List, NamedTuple, Optional, TypedDict
         import hashlib
         from urllib.parse import urljoin
+        from dataclasses import is_dataclass, asdict
+        from enum import Enum
 
         import stat
         import base64
@@ -431,6 +447,18 @@ def generate_jit_register_code(
 
         sys.stdout.reconfigure(line_buffering=True)
         sys.stderr.reconfigure(line_buffering=True)
+
+        def get_parameter_json_value(v):
+            if is_dataclass(v):
+                return asdict(v)
+            elif isinstance(v, Enum):
+                return v.value
+            elif isinstance(v, list):
+                return [get_parameter_json_value(x) for x in v]
+            elif isinstance(v, dict):
+                return {k: get_parameter_json_value(x) for k, x in v.items()}
+            else:
+                return v
 
         def check_exists_and_rename(old: Path, new: Path):
             if new.exists():
